Add in player object
Alright, this is a big commit, so I'll do a quick rundown of what it is:
This is a Blueprints based Pawn. It doesn't use UE4's physics, due to how it's locked to being vertical - although, if people want, I could change it to use it.
There's other issues with that (mostly just that the pawn bounces upon contact), which is why I'm using this method instead.
Currently, the Pawn relies on the static mesh to handle collissions, we're going to need to change that.
It has 360 degree movement along the X and Z axis. It rotates along the Y axis.
This is achieved through a line trace, and getting the normal of the colliding floor.
The blueprints are a bit of a mess, but I've commented/tidied up a bit.
It's state based, with different states having different properties. We can add more whenever, through PawnState.
The states are currently:
	Walking
	Running
	Falling
	Jumping